requirements.txt
numpy==2.3.4
PyQt6==6.10.0
PyOpenGL==3.1.10
PyOpenGL_accelerate==3.1.10
matplotlib>=3.0
# Add other project-specific dependencies here
qrcode
Pillow
